% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_md.R
\name{add_md}
\alias{add_md}
\title{Add markdown content to an XML object}
\usage{
add_md(body, md, where = 0L)
}
\arguments{
\item{body}{an XML object generated via {tinkr}}

\item{md}{a string of new markdown to insert}

\item{where}{the position in the markdown document to insert the new markdown}
}
\value{
a copy of the XML object with the markdown inserted.
}
\description{
Add markdown content to an XML object
}
\keyword{internal}
