% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{create_network}
\alias{create_network}
\title{Create igraph subnetwork}
\usage{
create_network(subLinks, nodes)
}
\arguments{
\item{subLinks}{a list of sublinks forming the network}

\item{nodes}{a master node list to pull information for the subnetwork}
}
\description{
Create igraph subnetwork
}
