% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{create_nodes}
\alias{create_nodes}
\title{Create a standardized tibble of nodes}
\usage{
create_nodes(vect, type, color)
}
\arguments{
\item{vect}{a vector of unique names for the nodes}

\item{type}{the type of nodes. This can be used for filtering}

\item{color}{what color should this node have in plots?}
}
\description{
Create a standardized tibble of nodes
}
