% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{different_first_layer}
\alias{different_first_layer}
\title{Find a first layer difference between two nodes}
\usage{
different_first_layer(wantNodes, notNodes, links)
}
\arguments{
\item{wantNodes}{a string vector of nodes of interest}

\item{notNodes}{a string vector of nodes whose first order connections we should not include}

\item{links}{a master list of links to pull information from}
}
\description{
Find a first layer difference between two nodes
}
