% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{find_links}
\alias{find_links}
\title{Find links attached to a node or multiple nodes of interest}
\usage{
find_links(nodes, links, direction = "both", type = NA)
}
\arguments{
\item{nodes}{a string vector of nodes}

\item{links}{a master list of links to filter nodes from}

\item{direction}{c('both', 'in', 'out') filter the links down by direction. 'in' edges only include target nodes and 'out' edges only include source nodes}

\item{type}{filter links by type}
}
\description{
Find links attached to a node or multiple nodes of interest
}
