% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{find_links_n}
\alias{find_links_n}
\title{Find links of greater than 1 depth}
\usage{
find_links_n(nodes, links, n, direction = "both", type = NA)
}
\arguments{
\item{nodes}{string vector of node or nodes to start search}

\item{links}{master list of links to search connections from}

\item{direction}{c('both', 'in', 'out') only find links of a particular direction}

\item{type}{only find links of a particular type}
}
\description{
Find links of greater than 1 depth
}
