% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{find_nodes}
\alias{find_nodes}
\title{Pull nodes out from a list of sublinks}
\usage{
find_nodes(nodes, links, type = NA)
}
\arguments{
\item{nodes}{a master list of nodes and their attributes}

\item{links}{a list of sublinks}

\item{type}{filter nodes down by type}
}
\description{
Pull nodes out from a list of sublinks
}
