% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{get_secondary_link_count}
\alias{get_secondary_link_count}
\title{How similiar are two disconnected nodes}
\usage{
get_secondary_link_count(query, links)
}
\arguments{
\item{query}{string vector of node or nodes of interest}

\item{links}{a master list of links to pull information from}
}
\description{
How similiar are two disconnected nodes
}
