% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_md.R
\name{insert_md}
\alias{insert_md}
\title{Insert markdown before or after a set of nodes}
\usage{
insert_md(body, md, nodes, where = "after", space = TRUE)
}
\arguments{
\item{body}{an XML object generated via {tinkr}}

\item{md}{markdown text to insert}

\item{nodes}{a character vector of an XPath expression OR an \code{xml_node} or
\code{xml_nodeset} object.}

\item{where}{the position in the markdown document to insert the new markdown}

\item{space}{when \code{TRUE} (default) inline nodes have a single space appended
or prepended to avoid the added markdown abutting text.}
}
\value{
a copy of the XML object with the translated markdown inserted
}
\description{
Insert markdown before or after a set of nodes
}
\note{
The markdown content must be of the same type as the XML nodes, either
inline or block content.
}
\keyword{internal}
