% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_md.R
\name{md_to_xml}
\alias{md_to_xml}
\title{Convert markdown to XML}
\usage{
md_to_xml(md)
}
\arguments{
\item{md}{a character vector of markdown text}
}
\value{
an XML nodeset of the markdown text
}
\description{
Convert markdown to XML
}
\examples{
if (requireNamespace("withr")) {

withr::with_namespace("tinkr", {
md_to_xml(c(
  "## This is a new section of markdown",
  "",
  "Each new element",
  "Is converted to a new line of markdown text",
  "",
  "```{r code-example, echo = FALSE}",
  "cat('code blocks work well here, too')",
  "```",
  "",
  "Neat, right?"
))
})

}
}
\keyword{internal}
