% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{move}
\alias{move}
\title{An imitation of Gremlin's out() and in() functions}
\usage{
move(vertices, direction, graph, edge_req = NULL)
}
\description{
An imitation of Gremlin's out() and in() functions
}
