% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-protection.R
\name{add_protected_ranges}
\alias{add_protected_ranges}
\alias{is_protected}
\alias{get_protected_ranges}
\alias{remove_protected_ranges}
\title{Handle protected ranges for a node}
\usage{
add_protected_ranges(node, start, end)

is_protected(node)

get_protected_ranges(node)

remove_protected_ranges(node)
}
\arguments{
\item{node}{an XML \verb{<text>} node.}

\item{start}{\verb{\[integer\]} a vector of starting indices of a set of ranges}

\item{end}{\verb{\[integer\]} a vector of ending indices that are paired with
\code{start}}
}
\value{
\itemize{
\item \code{add_protected_ranges()}: the modified node
\item \code{remove_protected_ranges()}: the modified node
\item \code{is_protected()}: \code{TRUE} if the node has protection attributes
\item \code{get_protected_ranges()} a list containing integer vectors \code{start} and
\code{end} if the node is protected, otherwise, it returns NULL
}
}
\description{
Handle protected ranges for a node
}
