% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-links.R
\name{resolve_anchor_links}
\alias{resolve_anchor_links}
\title{Resolve Reference-Style Links}
\usage{
resolve_anchor_links(body, txt, ns = md_ns())
}
\arguments{
\item{body}{an XML body}

\item{ns}{an the namespace that resolves the Markdown namespace}

\item{path}{path to the source file}

\item{encoding}{the encoding of the file. Defaults to UTF-8}
}
\description{
\href{https://www.markdownguide.org/basic-syntax/#reference-style-links}{Reference style links} are
a form of markdown syntax that reduces dupcliation and makes markdown more
readable. They come in two parts:
\enumerate{
\item The inline part that uses two pairs of square brackets where the second
pair of square brackets contains the reference for the anchor part of the
link. Example:\preformatted{[inline text describing link][link-reference]
}
\item The anchor part, which can be anywhere in the document, contains a pair
of square brackets followed by a colon and space with the link and
optionally the link title. Example:\preformatted{[link-reference]: https://docs.ropensci.org/tinkr/ 'documentation for tinkr'
}
}

Commonmark treats reference-style links as regular links, which can be a
pain when converting large documents. This function resolves these
links by reading in the source document, finding the reference-style links,
and adding them back at the end of the document with the 'anchor' attribute.
}
\examples{
f <- system.file("extdata", "link-test.md", package = "tinkr")
md <- yarn$new(f, sourcepos = TRUE, anchor_links = FALSE)
md$show()
lnks <- tinkr:::resolve_anchor_links(md$body, readLines(md$path))
md$body <- lnks
md$show()
}
\keyword{internal}
