% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-links.R
\name{resolve_links}
\alias{resolve_links}
\title{Resolve Reference-Style Links}
\usage{
resolve_links(body, path, ns = md_ns(), encoding = "UTF-8")
}
\arguments{
\item{body}{an XML body}

\item{path}{path to the source file}

\item{ns}{an the NS that resolves the md ns}

\item{encoding}{the encoding of the file. Defaults to UTF-8}
}
\description{
Commonmark treats reference-style links as regular links, which can be a
pain when converting large documents. This is an attempt at resolving these
links by reading in the source document, finding the relative links, and
adding them back with the 'asis' attribute.
}
\examples{
f <- system.file("extdata", "link-test.md", package = "tinkr")
md <- yarn$new(f, sourcepos = TRUE)
md$show()
lnks <- tinkr:::resolve_links(md$body, md$path)
md$body <- lnks
md$show()
}
\keyword{internal}
