% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{reverse_degree}
\alias{reverse_degree}
\title{same as degree() but with the first input being vertex vector}
\usage{
reverse_degree(v, graph, mode = "all")
}
\arguments{
\item{v}{vector of vertices}

\item{graph}{an igraph to analyze}

\item{mode}{direction of edges to calculate degree}
}
\description{
same as degree() but with the first input being vertex vector
}
