% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{reverse_ego}
\alias{reverse_ego}
\title{same as ego() but with the first input being a vertex vector}
\usage{
reverse_ego(nodes, graph, order = 1, mode = "all")
}
\arguments{
\item{nodes}{vertex of nodes of interest}

\item{graph}{to find nodes of interest}

\item{order}{retain nodes within n-ordered distance of nodes of interest}

\item{mode}{which direction of links to follow}
}
\description{
same as ego() but with the first input being a vertex vector
}
