% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{reverse_head_of}
\alias{reverse_head_of}
\title{same as head_of() but with the first input being an edge vector}
\usage{
reverse_head_of(es, graph)
}
\arguments{
\item{es}{vector of edges}

\item{graph}{an igraph to analyze}
}
\description{
same as head_of() but with the first input being an edge vector
}
