% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{reverse_induced_subgraph}
\alias{reverse_induced_subgraph}
\title{same as induced_subgraph() but with the first input being a vertex vector}
\usage{
reverse_induced_subgraph(nodes, graph)
}
\arguments{
\item{nodes}{a vector of nodes or node names}

\item{graph}{the igraph to filter}
}
\description{
same as induced_subgraph() but with the first input being a vertex vector
}
