% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{reverse_subgraph_edges}
\alias{reverse_subgraph_edges}
\title{same as subgraph.edges() but with the first input being an edge vector}
\usage{
reverse_subgraph_edges(edges, graph, delete.vertices = T)
}
\arguments{
\item{edges}{a vector of edges}

\item{graph}{the graph to filter}

\item{delete.vertices}{boolean fo whether or not to delete unconnected vertices}
}
\description{
same as subgraph.edges() but with the first input being an edge vector
}
