% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{save_as}
\alias{save_as}
\title{saves an object to the global environment to be referenced later does not change the input object and the pipe chain can continue}
\usage{
save_as(x, name)
}
\arguments{
\item{x}{an object to save}

\item{name}{the variable name to save the object o}
}
\description{
saves an object to the global environment to be referenced later does not change the input object and the pipe chain can continue
}
