% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{smooth_transition}
\alias{smooth_transition}
\title{changes the object being piped to a different object}
\usage{
smooth_transition(x, y)
}
\arguments{
\item{x}{the object to transition from}

\item{y}{the object to transition to}
}
\description{
changes the object being piped to a different object
}
