% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-splits.R
\name{split_protected_nodes}
\alias{split_protected_nodes}
\alias{join_split_nodes}
\title{Split and Join nodes that have been protected}
\usage{
split_protected_nodes(body)

join_split_nodes(body)
}
\arguments{
\item{body}{an \code{xml_document} class object}
}
\value{
\itemize{
\item \code{split_protected_nodes()} - a copy of \code{body} where text nodes with
protection are split into 'asis' and regular nodes.
\item \code{join_split_nodes()} - the \code{body} object where text nodes that have been
split (those with the \verb{@split-id} attribute) are joined into a single text
node.
}
}
\description{
Split and Join nodes that have been protected
}
\examples{
ex <- system.file("extdata", "math-example.md", package = "tinkr")
m <- tinkr::yarn$new(ex)
m$protect_math()
# protection gives us protected nodes
get_protected_nodes(m$body)
original_body <- m$body
# ---- splitting --------------------------------------------
# splitting transforms those nodes into split text nodes
split_body <- split_protected_nodes(m$body)
get_protected_nodes(split_body)
xml2::xml_find_all(split_body, ".//node()[@split-id]")
# The effect is the same
m$head(10)
m$body <- split_body
m$head(10)
# ---- joining ----------------------------------------------
# joining is done in place
join_split_nodes(m$body)
m$head(10)
get_protected_nodes(m$body)
get_protected_nodes(original_body)
}
