% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesheet.R
\name{stylesheet}
\alias{stylesheet}
\title{The {tinkr} stylesheet}
\usage{
stylesheet()
}
\value{
a single element character vector representing the path to the
stylesheet used by {tinkr}.
}
\description{
This function returns the path to the {tinkr} stylesheet
}
\examples{
tinkr::stylesheet()
}
