% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asis-nodes.R
\name{toss_broken_teeth}
\alias{toss_broken_teeth}
\title{Re-align mis-aligned pairs of logical vectors}
\usage{
toss_broken_teeth(no_end, no_start)
}
\arguments{
\item{no_end}{\link{logical} vector indicating broken math elements that
have no ending pair}

\item{no_start}{\link{logical} vector of the same length as \code{no_end} indicating
broken math elements that have no opening pair.}
}
\value{
a logical vector the same length as the inputs that indicates which
corresponding pairs of inputs are aligned.
}
\description{
Re-align mis-aligned pairs of logical vectors
}
\details{
The primary purpose of this function is to determine which nodes in a
nodelist have a math expression broken across a line. We do this by feeding
in two logical vectors: one for math beginning elements that have no end and
one for math end elements that have no beginning. This function returns TRUE
for each element that is a paired math expression and \code{FALSE} for each
element that is unpaired.

Imagine a zipper. Every tooth fits in a groove directly opposite. If there
is a mistake and there is an extra tooth on one side, the zipper gets stuck.
the solution is to remove that tooth to realign the zipper.

This function takes two logical vectors assuming the following:
\enumerate{
\item no_end starts with TRUE
\item no_start ends with TRUE
\item no_end and no_start are the same length
}

This function recursively slides down the length of the vectors and applies
these rules:
\enumerate{
\item (exit case) if there are fewer than 2 elements left in the inputs, then
we return a \code{FALSE} for each element.
\item if the current no_end value matches the value of the next no_start, then
these two are complementary parts of the zipper and are both marked \code{TRUE}
and the index advances by two.
\item otherwise a single \code{FALSE} is returned and the index advances by 1.
}
}
\examples{
no_starts <- c(FALSE, TRUE,  FALSE, FALSE, FALSE, FALSE, TRUE)
no_ends   <- c(TRUE,  FALSE, TRUE,  TRUE,  TRUE,  TRUE,  FALSE)
toss_broken_teeth(no_ends, no_starts)
}
\keyword{internal}
