% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protected.R
\name{get_protected}
\alias{get_protected}
\title{Get protected nodes}
\usage{
get_protected(body, type = NULL, ns = md_ns())
}
\arguments{
\item{body}{an \code{xml_document} object}

\item{type}{a character vector listing the protections to be included.
Defaults to \code{NULL}, which includes all protected nodes:
\itemize{
\item math: via the \code{protect_math()} function
\item curly: via the \code{protect_curly()} function
\item unescaped: via the \code{protect_unescaped()} function
}}

\item{ns}{the namespace of the document (defaults to \code{\link[=md_ns]{md_ns()}})}
}
\value{
an \code{xml_nodelist} object.
}
\description{
Get protected nodes
}
\examples{
path <- system.file("extdata", "basic-curly.md", package = "tinkr")
ex <- tinkr::yarn$new(path, sourcepos = TRUE)
# protect curly braces
ex$protect_curly()
# add math and protect it
ex$add_md(c("## math\n", 
  "$c^2 = a^2 + b^2$\n", 
  "$$",
  "\\\\sum_{i}^k = x_i + 1",
  "$$\n")
)
ex$protect_math()
# get protected now shows all the protected nodes
get_protected(ex$body)
get_protected(ex$body, c("math", "curly")) # only show the math and curly
}
