% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{isolate_nodes}
\alias{isolate_nodes}
\alias{provision_isolation}
\title{Isolate nodes in a document}
\usage{
isolate_nodes(nodelist, type = "context")

provision_isolation(nodelist)
}
\arguments{
\item{nodelist}{an object of class \code{xml_nodelist} OR \code{xml_node} OR a list of
either.}

\item{type}{a string of either "context" (default), "censor", or "list"}
}
\value{
a list of two elements:
\itemize{
\item doc: a copy of the document with the nodes isolated depending on the
context
\item key: a string used to tag nodes that are isolated via the \code{tnk-key}
attribute
}
}
\description{
Isolate nodes in a document
}
\details{
\code{isolate_nodes()} and \code{provision_isolation()} are the workhorses for the
\code{show} family of functions. These functions will create a copy of the
document with the nodes present in \code{nodelist} isolated.
\itemize{
\item \code{isolate_nodes()} provides a switch between specific modes:
\itemize{
\item "context" include the nodes within the block context of the document.
For example, if the nodelist contains links in headings, paragraphs, and
lists, those links will appear within these blocks. When \code{mark = TRUE},
ellipses \verb{[...]} will be added to indicate hidden content.
\item "censor" by default will replace all non-whitespace characters with a
censor character. This is controlled by \code{tinkr.censor.regex} and
\code{tinkr.censor.mark}
\item "list" creates a new document and copies over the nodes so they appear
as a list of paragraphs.
}
\item \code{provision_isolation()} uses \code{\link[xml2:xml_children]{xml2::xml_root()}} and \code{\link[xml2:xml_path]{xml2::xml_path()}} to
make a copy of the root document and then tag the corresponding nodes in
the nodelist so that we can filter on nodes that are not connected to
those present in the nodelist.
}
}
\keyword{internal}
