% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_list}
\alias{show_list}
\alias{show_bare}
\alias{show_context}
\alias{show_censor}
\title{Display a node or nodelist as markdown}
\usage{
show_list(nodelist)

show_bare(nodelist)

show_context(nodelist)

show_censor(nodelist)
}
\arguments{
\item{nodelist}{an object of class \code{xml_nodelist} OR \code{xml_node} OR a list of
either.}
}
\value{
a character vector, displayed to the screen.
}
\description{
When inspecting the results of an XPath query, displaying the text often
}
\examples{
path <- system.file("extdata", "example1.md", package = "tinkr")
y <- tinkr::yarn$new(path, sourcepos = TRUE)
items <- xml2::xml_find_all(y$body, ".//md:item", tinkr::md_ns())
links <- xml2::xml_find_all(y$body, ".//md:link", tinkr::md_ns())
code <- xml2::xml_find_all(y$body, ".//md:code", tinkr::md_ns())
blocks <- xml2::xml_find_all(y$body, ".//md:code_block", tinkr::md_ns())

# show a list of items 
show_list(links[1:10])
show_list(code[1:10])
show_list(blocks[1:2])

# show the items in their local structure
show_bare(items)
show_bare(links)

# show the items with context markers ([...]) in the structure of the document
show_context(links[20:31])
show_context(code[1:10])

# show the items in the full document censored:
show_censor(links)
# you can set the mark to censor by using the `tinkr.censor option`
options(tinkr.censor = ".")
show_censor(links)
}
