% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinkr-package.R
\docType{package}
\name{tinkr-package}
\alias{tinkr}
\alias{tinkr-package}
\title{tinkr: Cast '(R)Markdown' Files to 'XML' and Back Again}
\description{
Casts '(R)Markdown' files to 'XML' and back to allow in-memory editing via of 'markdown' elements via 'XPath' through the 'R6' class 'yarn'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/tinkr/}
  \item \url{https://github.com/ropensci/tinkr}
  \item Report bugs at \url{https://github.com/ropensci/tinkr/issues}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})

Authors:
\itemize{
  \item Maëlle Salmon \email{msmaellesalmon@gmail.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
  \item Jeroen Ooms
}

Other contributors:
\itemize{
  \item Nick Wellnhofer (Nick Wellnhofer wrote the XSLT stylesheet.) [copyright holder]
  \item rOpenSci (https://ropensci.org/) [funder]
  \item Peter Daengeli [contributor]
}

}
\keyword{internal}
