% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_xml.R
\name{to_xml}
\alias{to_xml}
\title{Transform file to XML}
\usage{
to_xml(path, encoding = "UTF-8")
}
\arguments{
\item{path}{Path to the file.}

\item{encoding}{Encoding to be used by readLines.}
}
\value{
A list containing the YAML of the file (yaml)
and its body (body) as XML.
}
\description{
Transform file to XML
}
\examples{
path <- system.file("extdata", "example1.md", package = "tinkr")
post_list <- to_xml(path)
names(post_list)
}
