% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_xml.R
\name{to_xml}
\alias{to_xml}
\title{Transform file to XML}
\usage{
to_xml(path, encoding = "UTF-8", sourcepos = FALSE)
}
\arguments{
\item{path}{Path to the file.}

\item{encoding}{Encoding to be used by readLines.}

\item{sourcepos}{passed to [commonmark::markdown_xml()]. If `TRUE`, the
source position of the file will be included as a "sourcepos" attribute.
Defaults to `FALSE`.}
}
\value{
A list containing the YAML of the file (yaml)
and its body (body) as XML.
}
\description{
Transform file to XML
}
\examples{
path <- system.file("extdata", "example1.md", package = "tinkr")
post_list <- to_xml(path)
names(post_list)
path2 <- system.file("extdata", "example2.Rmd", package = "tinkr")
post_list2 <- to_xml(path2)
post_list2
}
